ativ5-m8/requirements.txt
wheel
gradio
langchain
python-dotenv